requirements/dev.txt
amqp==2.6.1
asgiref==3.6.0
attrs==21.2.0
Babel==2.12.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.2
billiard==3.6.4.0
black==21.9b0
blinker==1.4
boto3==1.17.84
botocore==1.20.112
celery==4.4.7
certifi==2020.6.20
cffi==1.15.1
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==3.0.1
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.8
cssutils==2.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.7.0
dj-database-url==1.2.0
Django==3.2.12
django-3-jet==1.0.8
dj-database-url==1.2.0
django-activity-stream==0.10.0
django-auto-prefetching==0.1.10
django-celery-beat==2.0.0
django-cors-headers==3.7.0
django-dotenv==1.4.2
django-extensions==3.1.3
django-filter==2.4.0
django-flower==1.0.0
django-health-check==3.16.4
django-heroku==0.3.1
django-inlinecss==0.3.0
django-jinja==2.10.2
django-model-utils==4.1.1
django-money==2.0
django-nose==1.4.7
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
django-storages==1.11.1
django-summernote==0.8.11.6
django-timezone-field==4.2.3
django-unique-upload==0.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.20.0
easy-thumbnails==2.7.1
factory-boy==3.2.0
Faker==17.5.0
fasteners==0.14.1
filelock==3.9.0
future==0.18.2
gevent==22.10.2
greenlet==2.0.2
guake==3.8.5
gunicorn==20.1.0
httplib2==0.20.2
humanize==0.5.1
identify==2.5.18
idna==2.9
importlib-metadata==4.6.4
inflection==0.5.1
ipdb==0.13.8
ipython==7.24.0
itypes==1.2.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==0.10.0
jsonschema==3.2.0
keyring==23.5.0
kombu==4.6.11
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
librabbitmq==2.0.0
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==8.10.0
mypy-extensions==1.0.0
netifaces==0.11.0
newrelic==6.4.0.157
nodeenv==1.7.0
nose==1.3.7
oauthlib==3.2.0
olefile==0.46
openapi-codec==1.3.2
packaging==23.0
paramiko==2.9.3
parso==0.8.3
pathspec==0.11.0
pbr==5.8.0
pexpect==4.8.0
pickleshare==0.7.5
pika==1.3.1
Pillow==9.0.1
platformdirs==3.0.0
plumbum==1.8.1
pre-commit==2.15.0
prompt-toolkit==3.0.38
protobuf==3.12.4
psycopg2==2.9.5
psycopg2-binary==2.9.5
ptyprocess==0.7.0
py-moneyed==1.2
pycparser==2.21
Pygments==2.14.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynliner==0.8.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-crontab==2.7.1
python-dateutil==2.8.1
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.1
pyxdg==0.27
PyYAML==5.4.1
redis==3.5.3
regex==2022.10.31
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rpyc==5.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.4.2
SecretStorage==3.3.1
sentry-sdk==1.1.0
simplejson==3.18.3
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.4
sqlparse==0.4.3
texttable==1.6.4
toml==0.10.2
tomli==1.2.3
traitlets==5.9.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.5
vine==1.3.0
virtualenv==20.20.0
wadllib==1.3.6
wcwidth==0.2.6
websocket-client==0.32.0
whitenoise==6.4.0
xdg==5.0.0
zipp==1.0.0
zope.event==4.6
zope.interface==5.5.2
